+++ CHANGELOG.md
@@ -23,10 +23,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [v19.0.0-preview.4] - Unreleased
 
+> [Bot API 6.7](https://core.telegram.org/bots/api#april-21-2023) (April 21, 2023)
+
 > [Bot API 6.6](https://core.telegram.org/bots/api#march-9-2023) (March 9, 2023)
 
 ### Added
 
+- Type `InlineQueryResultsButton`
+- Type `SwitchInlineQueryChosenChat`
+- Type `BotName`
+- Property `WebAppName` to the class `WriteAccessAllowed`
+- Property `SwitchInlineQueryChosenChat` to the class `InlineKeyboardButton`
+- Property `ViaChatFolderInviteLink` to the class `ChatMemberUpdated`
+- Static method `InlineKeyboardButton.WithSwitchInlineQueryChosenChat`
+- Requests related to bot's name:
+  - `SetMyNameRequest`
+  - `GetMyNameRequest`
+- Methods related to bot's name:
+  - `ITelegramBotClient.SetMyNameAsync`
+  - `ITelegramBotClient.GetMyNameAsync`
 - New requests related to description:
   - `GetMyDescriptionRequest`
   - `SetMyDescriptionRequest`
@@ -69,6 +84,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Changed
 
+- Replace the parameters `switchPmText` and `switchPmParameter` of the method `AnswerInlineQueryAsync` with the parameter `button` of type `InlineQueryResultsButton`.
+- Replace the properties `SwitchPmText` and `SwitchPmParameter` of the class `AnswerInlineQueryRequest` with the property `Button` of type `InlineQueryResultsButton`.
 - Class `SetStickerSetThumbRequest` renamed to `SetStickerSetThumbnailRequest`
 - Class `CreateNewStickerSetRequest` is no longer abstract
 - Class `AddStickerToSetRequest` is no longer abstract